feat: proof of concept
can start server and client and pass mouse events to client when the server cursor hits a specific screen edge

currently everything is hard coded and there is no way it would work on another setup. Also the only way to get
control of the mouse back on teh server is to disconnect from the client.